Added the wind interpolation method for use in balloon propagation *created interpolation method
